+++ src/python_build_utils/cli_tools.py
@@ -16,6 +16,7 @@ import click
 from . import __version__
 from .collect_dep_modules import collect_dependencies
 from .collect_pyd_modules import collect_pyd_modules
+from .pyd2wheel import pyd2wheel
 from .remove_tarballs import remove_tarballs
 from .rename_wheel_files import rename_wheel_files
 
@@ -26,6 +27,7 @@ def cli() -> None:
     """A collection of CLI tools for Python build utilities."""
 
 
+cli.add_command(pyd2wheel)
 cli.add_command(collect_pyd_modules)
 cli.add_command(collect_dependencies)
 cli.add_command(rename_wheel_files)