+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.3] - 2025-04-15
+- Added new command line utility collect-dependencies in order to generate a list of all package dependencies
+- Added new command line utility collect-pyd-modules in order to generate a list of cythonized pyd dependencies:w
+
 ## [0.1.2] - 2025-02-22
 
 ### Added